requirements.txt
streamlit
Pillow
numpy
opencv-python
scikit-learn
tensorflow
